Add some tests for the Git service
Eventually the Git service will be split into smaller independent services,
but first I need a large number of tests to ensure nothing breaks!

Once all the tests are done the handlers can be updated to use the new services,
and the functionality can be pushed down to these services.