Introduce GitTempRepository for GitSplitSh
Move the sub-procedure to split a repository into a separate
service from which all operations are formed.

The main repository no longer holds the remotes for split repositories.
This reduces the risk that git update provides by pulling-in all tags from
split-repositories.

GitSplitSh pushes to these temporary repositories by using the local-URI instead.

Note: Unlike the old version this will "keep" the temp-repositories in user-temp
folder to speed-up the process. Previously this was removed after the command ended.