Store branches-alias in config
The branch-alias can now be stored in the repository configuration,
and it's now possible to specify an alias for other branches than the primary.

Storing the branch-alias in Git configuration is now deprecated and will be removed
in v2.0, this requires manual work and might become outdated among team members,
secondly this is a hack for Git as custom fields are not really supported.

Note: The composer value still takes precedence over the repository configuration.