Store pull-request metadata with git notes
Currently the metadata for a pull-request (number, category, authors, special labels)
is stored as part of the commit message.

In HuPKit v2.0 this will change completely to allow for customizing everything,
including a custom templates, and custom categories (among many other things).

To make upgrading easier the storage of pr-metadata is now *also* stored with Git notes
(ref: pr-metadata). The old (current) format and parsing is still used and will remain supported
in v2.0 (using a configuration flag), and support for commit-message based parsing will be (eventually)
removed in v3.0