Prevent branch switching with GitFileReader
The GitFileReader uses a temporary directory to clone to and read from,
this clone is always singular to reduce too much temp directories.

But when using the GitFileReader it's possible this might cause some side-effects
when one file reads from branch A but that file uses the GitFileReader to read from branch B,
which would cause a switch (without switching back) and file A not being able to load any other
files from branch A (as was expected).

Hopefully this will also fix a rather hard to find bug where configuration loading fails randomly